requirements.txt
firebase-rest-api
adafruit-circuitpython-si7021
board

